+++ src/R-read-png.c
@@ -168,6 +168,89 @@ spng_ctx *read_png_core(SEXP src_, FILE **fp, int rgba, int *fmt, int image_type
 }
 
 
+
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+// Read PNG as raw() vector
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+SEXP read_png_as_raw_(SEXP src_, SEXP rgba_, SEXP flags_) {
+  
+  FILE *fp = NULL;
+  int fmt   = SPNG_FMT_RGBA8;
+  uint8_t bits  = 8; 
+  int flags = asInteger(flags_);
+  uint32_t nchannels;
+  
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  // Create a context 
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  uint32_t width  = 0;
+  uint32_t height = 0;
+  size_t out_size = 0;
+  spng_ctx *ctx = read_png_core(src_, &fp, asInteger(rgba_), &fmt, R_IMAGE_ARRAY, &width, &height, &out_size, &bits, &nchannels);
+  
+  if (bits == 16) {
+    error("read_png_as_raw_(): 16 bit not handled yet");
+    // return read_png_as_array16_(src_, rgba_, flags_, avoid_transpose_);
+  }
+  
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  // 
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  int npixels = (int)(width * height);
+  nchannels = out_size / npixels;
+  
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  // Initialise memory into which the PNG will be decoded
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  unsigned char *decode_buf = (unsigned char *)malloc(out_size);
+  if (decode_buf == NULL) {
+    if (fp) fclose(fp);
+    spng_ctx_free(ctx);
+    error("Couldn't allocate PNG buffer");
+  }
+  
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  // Decode
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  int err = spng_decode_image(ctx, decode_buf, out_size, fmt, flags);
+  if (err) {
+    if (fp) fclose(fp);
+    free(decode_buf);
+    spng_ctx_free(ctx);
+    error("spng_decode_image() error: %s\n", spng_strerror(err));
+  }
+  
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  // Prep space for raw bytes
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  SEXP res_ = PROTECT(allocVector(RAWSXP, (R_xlen_t)out_size));
+  
+  unsigned char *res_ptr = RAW(res_);
+  unsigned char *buf_ptr = decode_buf;
+  
+  memcpy(res_ptr, buf_ptr, out_size);
+  
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  // Set attributes on result
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  setAttrib(res_, mkString("width") , ScalarInteger((int)width));
+  setAttrib(res_, mkString("height"), ScalarInteger((int)height));
+  setAttrib(res_, mkString("depth") , ScalarInteger((int)nchannels));
+  setAttrib(res_, mkString("bits" ) , ScalarInteger(8));
+
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  // Tidy and return
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  if (fp) fclose(fp);
+  spng_ctx_free(ctx);
+  free(decode_buf);
+  UNPROTECT(1);
+  return res_;
+}
+
+
+
+
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 // Read PNG as a nativeraster
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -832,6 +915,8 @@ SEXP read_png_(SEXP src_, SEXP type_, SEXP rgba_, SEXP flags_, SEXP avoid_transp
     return read_png_as_array_(src_, rgba_, flags_, avoid_transpose_);
   } else if (strcmp(image_type, "indexed") == 0) {
     return read_indexed_png_as_indexed_(src_, rgba_, flags_, avoid_transpose_);
+  } else if (strcmp(image_type, "raw") == 0) {
+    return read_png_as_raw_(src_, rgba_, flags_);
   }
   
   error("image type not understood: %s", image_type);